SF: Add retry to EventThread postEvent
When consumer BitTube is under heavy Event load, like constantly
updating use case providing VSYNC events, it is possible for Hotplug
events to fail sending across the socket. In such case, retry
mechanism is needed to send it again. We would like not to invoke
retry logic so we are also doubling the BitTube size to avoid
Event backlog in the first place.

Change-Id: Ief5cbd4ba662224a5595294e24f53bccbf65e103
CRs-Fixed: 2700045
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>