Reduce timeouts for occasionally-stalling sections
This change reduces the timeouts for INCIDENT REPORT and VISIBLE WINDOW
VIEWS to 20s and 5s. This is because go/bugreport-durations shows that
these sections occasionally stall to the full timeout (120s), but
otherwise finish under 10s and 1s. This change should reduce the
likelihood of these sections dramatically increasing the overall
duration of a bug report.

Test: Collect a bug report with 'adb bugreportz' to verify that the
      incident report and visible window views sections still exist.
Bug: 249522178
Change-Id: Ie2d20cc0192bff892fdffc5e9bce84feba848553
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>